Add unit tests config
