+++ package.json
@@ -9,7 +9,6 @@
     "build": "tsc -p tsconfig-build.json --skipLibCheck && cp -R src/protos lib && cp src/schema.graphql lib",
     "start": "yarn build && source ./exportLocal.sh && node lib/entrypoint/graphql.js | pino-pretty -c -l",
     "trigger": "yarn build && source ./exportLocal.sh && node lib/entrypoint/trigger.js | pino-pretty -c -l",
-    "test": "export JEST_JUNIT_OUTPUT_DIR=\"./artifacts\" && jest --ci --reporters=default --reporters=jest-junit --forceExit",
     "watch": "nodemon -V -e ts,graphql -w ./src -x yarn run start",
     "watch-trigger": "nodemon -V -e ts,graphql -w ./src -x yarn trigger",
     "kill-graphql": "kill $(lsof -t -i:4000)",
@@ -17,11 +16,8 @@
     "uninstall-charts": "yarn uninstall-charts-helm || : && sleep 15 && kubectl delete --all pvc",
     "add-charts": "export LOCAL=true && source ./exportLocal.sh && ./initLocalTest.sh",
     "forward-mongodb": "kubectl port-forward $(kubectl get pods -l app.kubernetes.io/name=mongodb | awk '/mongo/ {print $1}') 27017",
-    "test-local-restart": "./deletePods.sh && source ./exportLocal.sh && yarn run test-local",
     "restart-charts": "yarn use-minikube && yarn uninstall-charts || : && sleep 5 && yarn run add-charts",
     "use-minikube": "kubectl config use-context minikube",
-    "test-local-clean": "yarn run restart-charts && yarn jest --clearCache && yarn run test-local",
-    "test-local": "yarn use-minikube && source ./exportLocal.sh && yarn jest --runInBand --verbose $TEST | pino-pretty -c -l",
     "cron": "yarn build && source exportLocal.sh && node src/entrypoint/cron.js",
     "exporter": "yarn build && source exportLocal.sh && node src/entrypoint/exporter.js",
     "lnd-connection": "source exportLocal.sh && echo $MACAROON | base64 -d > ./macaroon && echo $TLS | base64 -d > ./tls.cert && lncli --network regtest --macaroonpath=./macaroon --tlscertpath=./tls.cert --rpcserver=${LNDDNS}:${LNDRPCPORT} getinfo",
@@ -29,7 +25,13 @@
     "postman": "yarn node src/postman/newman.ts",
     "bitcoin-cli": "source ./exportLocal.sh && kubectl exec bitcoind-0 -- bitcoin-cli -rpcuser=rpcuser -rpcpassword=$BITCOINDRPCPASS -regtest",
     "update-postman": "curl --location --request GET \"https://api.getpostman.com/collections/4e3fee88-7267-44e1-a316-b81e90b0bd2b\" --header \"X-Api-Key: $POSTMANAPIKEY\" | jq . > src/postman/graphql.postman_collection.json",
-    "autocannon-rn-no-auth": "yarn autocannon https://graphql.testnet.galoy.io:443/graphql -m POST -H Content-Type=application/json -b '{\"query\":\"query gql_query_anonymous {\n  prices {\n    __typename\n    id\n    o\n  }\n  earnList {\n    __typename\n    id\n    value\n  }\n  maps {\n    __typename\n    id\n    title\n    coordinate {\n        __typename\n        latitude\n        longitude\n    }\n  }\n}\",\"variables\":{}}' -c 20 -d 10"
+    "autocannon-rn-no-auth": "yarn autocannon https://graphql.testnet.galoy.io:443/graphql -m POST -H Content-Type=application/json -b '{\"query\":\"query gql_query_anonymous {\n  prices {\n    __typename\n    id\n    o\n  }\n  earnList {\n    __typename\n    id\n    value\n  }\n  maps {\n    __typename\n    id\n    title\n    coordinate {\n        __typename\n        latitude\n        longitude\n    }\n  }\n}\",\"variables\":{}}' -c 20 -d 10",
+    "test": "export JEST_JUNIT_OUTPUT_DIR=\"./artifacts\" && jest --ci --reporters=default --reporters=jest-junit --forceExit",
+    "test-local": "yarn use-minikube && source ./exportLocal.sh && yarn jest --runInBand --verbose $TEST | pino-pretty -c -l",
+    "test-local-restart": "./deletePods.sh && source ./exportLocal.sh && yarn run test-local",
+    "test-local-clean": "yarn run restart-charts && yarn jest --clearCache && yarn run test-local",
+    "test:unit": "jest --config ./test/jest-unit.config.js --bail --verbose | yarn pino-pretty -c -l",
+    "test:integration": ". ./.envrc && jest --config ./test/jest-integration.config.js --forceExit --bail --runInBand --verbose | yarn pino-pretty -c -l"
   },
   "engines": {
     "node": "14"
