--- test/integration/09e-yaml.spec.ts
@@ -1,11 +1,0 @@
-import { defaultConfig } from "src/config"
-
-it("test", async () => {
-  try {
-    console.log(defaultConfig)
-    expect(defaultConfig).toHaveProperty("hedging")
-    expect(defaultConfig).toHaveProperty("name")
-  } catch (e) {
-    console.log(e)
-  }
-})
