+++ test/integration/09g-utils.spec.ts
@@ -1,18 +1,8 @@
-import { btc2sat, sat2btc, isInvoiceAlreadyPaidError } from "src/utils"
+import { isInvoiceAlreadyPaidError } from "src/utils"
 import { lndOutside1, lndOutside2 } from "./helper"
 import { createInvoice, pay } from "lightning"
 
 jest.mock("src/realtimePrice", () => require("../mocks/realtimePrice"))
-
-it("btc2sat", async () => {
-  const BTC = 1.2
-  expect(btc2sat(BTC)).toEqual(120000000)
-})
-
-it("sat2btc", async () => {
-  const sat = 120000000
-  expect(sat2btc(sat)).toEqual(1.2)
-})
 
 it("decodes lnservice error correctly", async () => {
   const { request } = await createInvoice({ lnd: lndOutside2, tokens: 1000 })
