+++ test/jest-unit.config.js
@@ -1,13 +1,13 @@
 module.exports = {
   moduleFileExtensions: ["js", "json", "ts"],
-  rootDir: ".",
-  roots: ["<rootDir>/test"],
+  rootDir: "../",
+  roots: ["<rootDir>/test/unit"],
   transform: {
     "^.+\\.(ts)$": "ts-jest",
   },
   testRegex: ".*\\.spec\\.ts$",
-  testSequencer: "./test/jest-test-sequencer.js",
-  setupFilesAfterEnv: ["./test/jest.setup.js"],
+  testSequencer: "<rootDir>/test/jest-test-sequencer.js",
+  setupFilesAfterEnv: ["<rootDir>/test/jest.setup.js"],
   testEnvironment: "node",
   moduleNameMapper: {
     "^src/(.*)$": "<rootDir>/src/$1",
